showing volume bar on volume change
